Centre the timestamps in the frame
Rather then having them in the top left corner.  We do this because the
Raspberry Pi is not fast enough to keep up with drawing frames at 720p60,
but for smaller resolutions it will center the video.  By centering the
timestamps we can drop the resolution while the timestamps will stay in
the same location.